Fix date types

+++ migrations/0000_same_newton_destine.sql
@@ -2,7 +2,7 @@ CREATE TABLE `tasks` (
 	`id` integer PRIMARY KEY AUTOINCREMENT NOT NULL,
 	`name` text NOT NULL,
 	`isComplete` integer DEFAULT false NOT NULL,
-	`createdAt` numeric DEFAULT (CURRENT_TIMESTAMP) NOT NULL
+	`createdAt` integer DEFAULT (CURRENT_TIMESTAMP) NOT NULL
 );
 --> statement-breakpoint
 CREATE TABLE `users` (
@@ -11,7 +11,7 @@ CREATE TABLE `users` (
 	`email` text NOT NULL,
 	`password` text NOT NULL,
 	`permissions` blob DEFAULT '[]' NOT NULL,
-	`createdAt` numeric DEFAULT (CURRENT_TIMESTAMP) NOT NULL
+	`createdAt` integer DEFAULT (CURRENT_TIMESTAMP) NOT NULL
 );
 --> statement-breakpoint
 CREATE UNIQUE INDEX `users_email_unique` ON `users` (`email`);
